Abstract out ModelStub class
